Simplify settings and persistence
